feat: ТЗ 16. Добавил тесты

+++ server/src/main/java/ru/practicum/shareit/user/service/UserServiceImpl.java
@@ -64,6 +64,10 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void delete(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new NotFoundException("User with id " + id + " not found");
+        }
+
         userRepository.deleteById(id);
     }
 }
